backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
redis
pydantic[email]
pydantic-settings
email-validator
weasyprint
celery
boto3
sentry-sdk
slowapi
